Separated login functions from the main activity
