+++ app/src/main/java/com/example/firebaseauth/MainActivity.java
@@ -1,8 +1,6 @@
 package com.example.firebaseauth;
 
-import android.content.Intent;
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.GridLayoutManager;
@@ -17,23 +15,13 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.example.firebaseauth.ADS.AdsRecyclerAdapter;
-import com.example.firebaseauth.LOCATION.LocationActivity;
-import com.facebook.FacebookSdk;
 import com.facebook.appevents.AppEventsLogger;
 import com.firebase.ui.auth.AuthUI;
-import com.firebase.ui.auth.IdpResponse;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-
-import java.util.Arrays;
-import java.util.List;
 
 public class MainActivity extends AppCompatActivity {
-    private static final int MY_REQUEST_CODE = 7117;
-    List<AuthUI.IdpConfig> providers;
     Button btn_sign_out;
     TextView headerTitle,subHeaderTitle;
     ImageView ic_cards;
@@ -59,15 +47,12 @@ public class MainActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
 
-        FacebookSdk.sdkInitialize(getApplicationContext());
         AppEventsLogger.activateApp(this);
         Animation ttb = AnimationUtils.loadAnimation(this,R.anim.ttb);
         Animation stb= AnimationUtils.loadAnimation(this,R.anim.scaletobig);
         //Animation button_anim= AnimationUtils.loadAnimation(this,R.anim.btn_nxt_btt);
-
-
-        setContentView(R.layout.activity_main);
 
         recyclerView = findViewById(R.id.activity_main_recycler_view);
         layoutManager = new GridLayoutManager(this,3);
@@ -94,8 +79,6 @@ public class MainActivity extends AppCompatActivity {
                             @Override
                             public void onComplete(@NonNull Task<Void> task) {
                                 btn_sign_out.setEnabled(false);
-                                showSignInOptions();
-
                             }
                         }).addOnFailureListener(new OnFailureListener() {
                     @Override
@@ -106,80 +89,37 @@ public class MainActivity extends AppCompatActivity {
             }
         });
 
-        providers = Arrays.asList(
-                new AuthUI.IdpConfig.EmailBuilder().build(),
-                new AuthUI.IdpConfig.PhoneBuilder().build(),
-                new AuthUI.IdpConfig.FacebookBuilder().build(),
-                new AuthUI.IdpConfig.GoogleBuilder().build()
-        );
-        headerTitle.startAnimation(ttb);
-        subHeaderTitle.startAnimation(ttb);
-        ic_cards.startAnimation(stb);
+
+         headerTitle.startAnimation(ttb);
+         subHeaderTitle.startAnimation(ttb);
+         ic_cards.startAnimation(stb);
 
         new android.os.Handler().postDelayed(
                 new Runnable() {
                     public void run() {
-                        showSignInOptions();
                         recyclerView.setAdapter(asdAdapter);
                     }
                 },
                 4000);
     }
 
-    private void showSignInOptions() {
-        mainLayout.setVisibility(LinearLayout.GONE);
-        startActivityForResult(
-                AuthUI.getInstance().createSignInIntentBuilder()
-                .setAvailableProviders(providers)
-                .setTheme(R.style.Mytheme)
-                .build(),MY_REQUEST_CODE
+//    @Override
+//    public void onBackPressed() {
+//        AuthUI.getInstance()
+//                .signOut(MainActivity.this)
+//                .addOnCompleteListener(new OnCompleteListener<Void>() {
+//                    @Override
+//                    public void onComplete(@NonNull Task<Void> task) {
+//                        btn_sign_out.setEnabled(false);
+//
+//                    }
+//                }).addOnFailureListener(new OnFailureListener() {
+//            @Override
+//            public void onFailure(@NonNull Exception e) {
+//                Toast.makeText(MainActivity.this,""+e.getMessage(),Toast.LENGTH_SHORT).show();
+//            }
+//        });
+//
+//    }
 
-        );
-    }
-    @Override
-    public void onBackPressed() {
-        AuthUI.getInstance()
-                .signOut(MainActivity.this)
-                .addOnCompleteListener(new OnCompleteListener<Void>() {
-                    @Override
-                    public void onComplete(@NonNull Task<Void> task) {
-                        btn_sign_out.setEnabled(false);
-                        showSignInOptions();
-
-                    }
-                }).addOnFailureListener(new OnFailureListener() {
-            @Override
-            public void onFailure(@NonNull Exception e) {
-                Toast.makeText(MainActivity.this,""+e.getMessage(),Toast.LENGTH_SHORT).show();
-            }
-        });
-
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        if(requestCode==MY_REQUEST_CODE){
-            IdpResponse idpResponse= IdpResponse.fromResultIntent(data);
-
-            if(resultCode==RESULT_OK){
-                FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
-                Toast.makeText(this,""+user.getEmail(),Toast.LENGTH_SHORT).show();
-                btn_sign_out.setEnabled(true);
-                new android.os.Handler().postDelayed(
-                        new Runnable() {
-                            public void run() {
-                                Intent intent = new Intent(MainActivity.this, LocationActivity.class);
-                                startActivity(intent);
-
-                            }
-                        },
-                        5000);
-
-            }else{
-                Toast.makeText(this,"Bye Bye",Toast.LENGTH_SHORT).show();
-                finish();
-            }
-        }
-    }
 }
