+++ app/src/main/res/layout/activity_login.xml
@@ -1,0 +1,80 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="#fff"
+    tools:context=".LOGIN.Login">
+
+    <TextView
+        android:id="@+id/title"
+        android:layout_width="173dp"
+        android:layout_height="wrap_content"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentTop="true"
+        android:layout_alignParentEnd="true"
+        android:layout_centerInParent="true"
+        android:layout_marginStart="119dp"
+        android:layout_marginTop="337dp"
+        android:layout_marginEnd="118dp"
+        android:fontFamily="@font/mmedium"
+        android:text="@string/shopit"
+        android:textAlignment="center"
+        android:textColor="#63b1ff"
+        android:textSize="40sp" />
+
+    <ImageView
+        android:id="@+id/logo"
+        android:layout_width="96dp"
+        android:layout_height="95dp"
+        android:layout_below="@+id/title"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentTop="true"
+        android:layout_marginStart="7dp"
+        android:layout_marginLeft="7dp"
+        android:layout_marginTop="7dp"
+        android:contentDescription="@string/todo"
+        android:src="@drawable/ic_cards" />
+
+    <TextView
+        android:id="@+id/activity_main_header_sub_title"
+        android:layout_width="230dp"
+        android:layout_height="86dp"
+        android:layout_below="@+id/logo"
+        android:layout_centerHorizontal="true"
+        android:layout_marginStart="88dp"
+        android:layout_marginLeft="88dp"
+        android:layout_marginTop="326dp"
+        android:fontFamily="@font/mlight"
+        android:text="@string/find_a_supermaket_near_look_for_the_best_price_and_shop"
+        android:textAlignment="center"
+        android:textSize="16sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
+
+    <ImageView
+        android:id="@+id/imageView"
+        android:layout_width="198dp"
+        android:layout_height="192dp"
+        android:layout_alignParentBottom="true"
+        android:layout_marginTop="499dp"
+        android:layout_marginBottom="17dp"
+        android:contentDescription="@string/todo"
+        app:srcCompat="@drawable/shopper" />
+
+    <ImageView
+        android:id="@+id/imageView2"
+        android:layout_width="203dp"
+        android:layout_height="198dp"
+        android:layout_alignParentTop="true"
+        android:layout_alignParentEnd="true"
+        android:layout_marginTop="106dp"
+        android:layout_marginEnd="0dp"
+        android:contentDescription="@string/todo"
+        app:srcCompat="@drawable/deliver" />
+
+</RelativeLayout>
