+++ app/src/main/res/layout/activity_main.xml
@@ -27,7 +27,7 @@
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="find a supermaket near, look for the best price and shop"
+            android:text="Find a Supermaket near, look for the best price and shop"
             android:fontFamily="@font/mlight"
             android:textSize="16dp"
             app:layout_constraintLeft_toLeftOf="parent"
