+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,15 @@
 <resources>
-    <string name="app_name">FireBaseAuth</string>
+    <string name="app_name">Shopit</string>
     <string name="facebook_application_id">1107418436112876</string>
     <string name="fb_login_protocol_scheme">fb1107418436112876</string>
     <string name="location_prompt">Choose a country</string>
+
+    <string name="title_activity_location">LocationActivity</string>
+
+    <string name="title_activity_login">Login</string>
+    <string name="shopit">Shopit</string>
+    <string name="find_a_supermaket_near_look_for_the_best_price_and_shop">Find a supermaket near, look for the best price and shop</string>
+    <string name="todo">TODO</string>
 
     <string-array name="location_arrays">
         <item>Limuru</item>
@@ -15,14 +22,12 @@
         <item>Kericho</item>
         <item>Ruaka</item>
     </string-array>
-
     <string-array name="Limuru">
         <item>Naivas</item>
         <item>Clean Shelf</item>
         <item>Rongai</item>
         <item>Limumatt</item>
     </string-array>
-
     <string-array name="Kiambu">
         <item>Naivas</item>
         <item>Clean Shelf</item>
@@ -58,6 +63,7 @@
         <item>Tuskys</item>
         <item>QuickMart</item>
     </string-array>
+
     <string-array name="Lamu">
         <item>Clean Shelf</item>
         <item>Tuskys</item>
@@ -66,6 +72,4 @@
         <item>Naivas</item>
         <item>Tuskys</item>
     </string-array>
-
-    <string name="title_activity_location">LocationActivity</string>
 </resources>
